requirements.txt
beautifulsoup4==4.3.2
Django==1.5
django-debug-toolbar==1.2.2
django-registration==1.0
django-wysiwyg==0.7.0
simplejson==3.6.5
six==1.9.0
sqlparse==0.1.14
